PET-300 refactor : 소셜 로그인 시 이메일 변경에 대한 처리 전략 정리 후 테스트 주석 처리
